backend/requirements.txt
fastapi
uvicorn[standard]
python-whois
requests
langchain
langchain-google-genai
langchain-openai
python-dotenv
pydantic
